move menu builder code out of manager.js
